Adds kernel packages for Focal
Reusing the xenial packages for Focal so that we have a baseline to work
from. We may end up shipping 5.x kernels soon, but we're still
working on building out Focal support.

What's missing here is the "securedrop-grsec" metapackage for Focal.
We don't have a unique version string for the metapackage, including a
platform such as +xenial or +focal, so we'll have to add some logic
to support the package.