Add template-consolidation component for buster
Using a dedicated "component" so that we can isolate ongoing development
specifically on template consolidation work [0] so that we can evaluate
upgrade behavior. While it's frankly acceptable to break "apt-test"
packages temporarily, we need the ability to test before/after states to
evaluate behavior during upgrades. Thus the separate component.

Requires updates to the CD logic that runs on apt-test.