Fix double signing on pushes
GHA is currently pushing two commits after new debs are committed
because the output is not fully idempotent. Running build.py always
modifies the repo/db/checksums.db file, which we already knew was a
problem from securedrop-apt-prod's CI.

So the fix is to run build.py, if the only thing it touched is
checksums.db, reset it and move on. The update metadata export step will
get skipped because nothing has changed, and no new commit will be
created.

Refs #267.