Fix update-tor workflow
Because the database is now committed to Git, the `reprepro update`
command is smarter and won't download the upstream debs if they're
already present in the repository.

This is generally good, but we want copies in our core/{focal,noble}
folders. Have reprepro update the copy in repo/public/pool, copy the
latest version of those debs over to our core/ folders. Git will commit
just the new core/ debs, and the next commit will trigger re-signing.

Refs #267.